requirements.txt
absl-py==1.2.0
astunparse==1.6.3
beautifulsoup4==4.11.1
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
dash==2.3.1
dash-canvas==0.1.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.0
Flask-Compress==1.12
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.21.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
MouseInfo==0.1.3
networkx==2.8.5
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
Pillow==9.2.0
plotly==5.9.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc==8.5
pyobjc-core==8.5
pyobjc-framework-Accessibility==8.5
pyobjc-framework-Accounts==8.5
pyobjc-framework-AddressBook==8.5
pyobjc-framework-AdServices==8.5
pyobjc-framework-AdSupport==8.5
pyobjc-framework-AppleScriptKit==8.5
pyobjc-framework-AppleScriptObjC==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-AppTrackingTransparency==8.5
pyobjc-framework-AudioVideoBridging==8.5
pyobjc-framework-AuthenticationServices==8.5
pyobjc-framework-AutomaticAssessmentConfiguration==8.5
pyobjc-framework-Automator==8.5
pyobjc-framework-AVFoundation==8.5
pyobjc-framework-AVKit==8.5
pyobjc-framework-BusinessChat==8.5
pyobjc-framework-CalendarStore==8.5
pyobjc-framework-CallKit==8.5
pyobjc-framework-CFNetwork==8.5
pyobjc-framework-ClassKit==8.5
pyobjc-framework-CloudKit==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Collaboration==8.5
pyobjc-framework-ColorSync==8.5
pyobjc-framework-Contacts==8.5
pyobjc-framework-ContactsUI==8.5
pyobjc-framework-CoreAudio==8.5
pyobjc-framework-CoreAudioKit==8.5
pyobjc-framework-CoreBluetooth==8.5
pyobjc-framework-CoreData==8.5
pyobjc-framework-CoreHaptics==8.5
pyobjc-framework-CoreLocation==8.5
pyobjc-framework-CoreMedia==8.5
pyobjc-framework-CoreMediaIO==8.5
pyobjc-framework-CoreMIDI==8.5
pyobjc-framework-CoreML==8.5
pyobjc-framework-CoreMotion==8.5
pyobjc-framework-CoreServices==8.5
pyobjc-framework-CoreSpotlight==8.5
pyobjc-framework-CoreText==8.5
pyobjc-framework-CoreWLAN==8.5
pyobjc-framework-CryptoTokenKit==8.5
pyobjc-framework-DeviceCheck==8.5
pyobjc-framework-DictionaryServices==8.5
pyobjc-framework-DiscRecording==8.5
pyobjc-framework-DiscRecordingUI==8.5
pyobjc-framework-DiskArbitration==8.5
pyobjc-framework-DVDPlayback==8.5
pyobjc-framework-EventKit==8.5
pyobjc-framework-ExceptionHandling==8.5
pyobjc-framework-ExecutionPolicy==8.5
pyobjc-framework-ExternalAccessory==8.5
pyobjc-framework-FileProvider==8.5
pyobjc-framework-FileProviderUI==8.5
pyobjc-framework-FinderSync==8.5
pyobjc-framework-FSEvents==8.5
pyobjc-framework-GameCenter==8.5
pyobjc-framework-GameController==8.5
pyobjc-framework-GameKit==8.5
pyobjc-framework-GameplayKit==8.5
pyobjc-framework-ImageCaptureCore==8.5
pyobjc-framework-IMServicePlugIn==8.5
pyobjc-framework-InputMethodKit==8.5
pyobjc-framework-InstallerPlugins==8.5
pyobjc-framework-InstantMessage==8.5
pyobjc-framework-Intents==8.5
pyobjc-framework-IOSurface==8.5
pyobjc-framework-iTunesLibrary==8.5
pyobjc-framework-KernelManagement==8.5
pyobjc-framework-LatentSemanticMapping==8.5
pyobjc-framework-LaunchServices==8.5
pyobjc-framework-libdispatch==8.5
pyobjc-framework-LinkPresentation==8.5
pyobjc-framework-LocalAuthentication==8.5
pyobjc-framework-MapKit==8.5
pyobjc-framework-MediaAccessibility==8.5
pyobjc-framework-MediaLibrary==8.5
pyobjc-framework-MediaPlayer==8.5
pyobjc-framework-MediaToolbox==8.5
pyobjc-framework-Metal==8.5
pyobjc-framework-MetalKit==8.5
pyobjc-framework-MetalPerformanceShaders==8.5
pyobjc-framework-MetalPerformanceShadersGraph==8.5
pyobjc-framework-MLCompute==8.5
pyobjc-framework-ModelIO==8.5
pyobjc-framework-MultipeerConnectivity==8.5
pyobjc-framework-NaturalLanguage==8.5
pyobjc-framework-NetFS==8.5
pyobjc-framework-Network==8.5
pyobjc-framework-NetworkExtension==8.5
pyobjc-framework-NotificationCenter==8.5
pyobjc-framework-OpenDirectory==8.5
pyobjc-framework-OSAKit==8.5
pyobjc-framework-OSLog==8.5
pyobjc-framework-PassKit==8.5
pyobjc-framework-PencilKit==8.5
pyobjc-framework-Photos==8.5
pyobjc-framework-PhotosUI==8.5
pyobjc-framework-PreferencePanes==8.5
pyobjc-framework-PushKit==8.5
pyobjc-framework-Quartz==8.5
pyobjc-framework-QuickLookThumbnailing==8.5
pyobjc-framework-ReplayKit==8.5
pyobjc-framework-SafariServices==8.5
pyobjc-framework-SceneKit==8.5
pyobjc-framework-ScreenSaver==8.5
pyobjc-framework-ScreenTime==8.5
pyobjc-framework-ScriptingBridge==8.5
pyobjc-framework-SearchKit==8.5
pyobjc-framework-Security==8.5
pyobjc-framework-SecurityFoundation==8.5
pyobjc-framework-SecurityInterface==8.5
pyobjc-framework-ServiceManagement==8.5
pyobjc-framework-Social==8.5
pyobjc-framework-SoundAnalysis==8.5
pyobjc-framework-Speech==8.5
pyobjc-framework-SpriteKit==8.5
pyobjc-framework-StoreKit==8.5
pyobjc-framework-SyncServices==8.5
pyobjc-framework-SystemConfiguration==8.5
pyobjc-framework-SystemExtensions==8.5
pyobjc-framework-UniformTypeIdentifiers==8.5
pyobjc-framework-UserNotifications==8.5
pyobjc-framework-UserNotificationsUI==8.5
pyobjc-framework-VideoSubscriberAccount==8.5
pyobjc-framework-VideoToolbox==8.5
pyobjc-framework-Virtualization==8.5
pyobjc-framework-Vision==8.5
pyobjc-framework-WebKit==8.5
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
PyWavelets==1.3.0
pywhatkit==5.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.2
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.9.0
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.7.31
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
wikipedia==1.4.0
wrapt==1.14.1
zipp==3.8.1
